safari-600.7-branch/LayoutTests/http/tests/navigation/redirect-to-random-url-versus-memory-cache-expected.txt

ALERT: Success: A different page was loaded at least once.
This test clicks a link that loads a URL into an iframe.
The URL redirects to a different, random URL every time the link is clicked.
When main resources are in the memory cache they are keyed off the original URL, not the post-redirect URL.
In the case where the redirect takes us to a different URL than the CachedResource we're revalidating, we need to abort the revalidate and issue an unconditional request.

This test makes sure that trying the random redirect multiple times results in at least one different load result.
Since the nature of this test is "random redirects", it runs enough multiple iterations to make it very likely we'll get at least one different redirect.
Click for random redirect
